refactor: unify PersonalityForm with create/edit modes
