--- .trellis/p/P-implement-personalities/e/E-ui-components-and-user/f/F-remove-localstorage-and/F-remove-localstorage-and.md
@@ -1,185 +1,0 @@
----
-id: F-remove-localstorage-and
-title: Remove localStorage and Finalize Integration
-status: open
-priority: medium
-parent: E-ui-components-and-user
-prerequisites:
-  - F-implement-user-interactions
-affectedFiles: {}
-log: []
-schema: v1.0
-childrenIds: []
-created: 2025-08-17T14:18:51.250Z
-updated: 2025-08-17T14:18:51.250Z
----
-
-# Remove localStorage and Finalize Store Integration
-
-## Purpose and Goals
-
-Complete the UI refactoring by removing all localStorage usage, finalizing the integration with the personalities store, and ensuring all CRUD operations work correctly with the file-based persistence layer. This is the final cleanup and integration feature.
-
-## Key Components to Implement
-
-### localStorage Removal
-
-- Search and remove ALL localStorage references
-- Remove draft personality storage code
-- Remove any localStorage utility functions
-- Clean up unused imports and dependencies
-- Remove localStorage keys from constants
-
-### Store Integration Finalization
-
-- Ensure all CRUD operations use store
-- Verify store initialization in PersonalitiesProvider
-- Confirm file persistence is working
-- Add proper error recovery
-- Implement retry logic for failed operations
-
-### Code Cleanup
-
-- Remove unused components (SavedPersonalitiesTab, etc.)
-- Delete obsolete utility functions
-- Clean up commented code
-- Update imports and exports
-- Remove unused props and state
-
-### Final Integration Testing
-
-- Verify create personality saves to file
-- Verify edit personality updates file
-- Verify delete personality updates file
-- Confirm personalities load on app restart
-- Test error scenarios
-
-## Detailed Acceptance Criteria
-
-### localStorage Cleanup
-
-- [ ] No localStorage.getItem() calls remain
-- [ ] No localStorage.setItem() calls remain
-- [ ] No localStorage.removeItem() calls remain
-- [ ] No localStorage key constants remain
-- [ ] No draft-related code remains
-- [ ] No auto-save functionality remains
-
-### Store Operations
-
-- [ ] Create personality uses store.createPersonality()
-- [ ] Update personality uses store.updatePersonality()
-- [ ] Delete personality uses store.deletePersonality()
-- [ ] List renders from store.personalities
-- [ ] Loading states from store.isLoading
-- [ ] Error states from store.error
-
-### File Persistence Verification
-
-- [ ] personalities.json created in user data directory
-- [ ] File updates on every change
-- [ ] Data persists across app restarts
-- [ ] Default personalities load on first run
-- [ ] File corruption handled gracefully
-
-### Component Cleanup
-
-- [ ] SavedPersonalitiesTab component deleted
-- [ ] Old CreatePersonalityForm deleted (if separate from PersonalityForm)
-- [ ] Unused tab-related components removed
-- [ ] Obsolete utility functions removed
-- [ ] Dead code eliminated
-- [ ] Unused imports cleaned up
-
-### Final Quality Checks
-
-- [ ] TypeScript compilation passes
-- [ ] No console errors or warnings
-- [ ] All props properly typed
-- [ ] Event handlers properly memoized
-- [ ] No memory leaks
-- [ ] Proper cleanup in useEffect hooks
-
-## Implementation Guidance
-
-### localStorage Search Pattern
-
-```bash
-# Search for all localStorage usage
-grep -r "localStorage" apps/desktop/src/components/settings/personalities/
-grep -r "draft" apps/desktop/src/components/settings/personalities/
-```
-
-### Store Connection Pattern
-
-```tsx
-const PersonalitiesSection = () => {
-  const {
-    personalities,
-    isLoading,
-    error,
-    createPersonality,
-    updatePersonality,
-    deletePersonality,
-    clearError,
-    retryLastOperation,
-  } = usePersonalitiesStore();
-
-  // No localStorage hooks
-  // No draft state
-  // Direct store operations only
-};
-```
-
-### Cleanup Checklist
-
-1. Remove localStorage utility imports
-2. Delete draft-related state and effects
-3. Remove auto-save timers/intervals
-4. Clean up draft-related UI elements
-5. Delete unused component files
-6. Update barrel exports
-
-### Verification Steps
-
-```typescript
-// After changes, verify:
-// 1. Create a personality -> Check personalities.json
-// 2. Edit a personality -> Verify file updates
-// 3. Delete a personality -> Confirm removal from file
-// 4. Restart app -> Personalities still present
-// 5. Delete personalities.json -> Defaults load
-```
-
-## Testing Requirements
-
-- No localStorage references in codebase
-- All CRUD operations work correctly
-- File persistence verified
-- App restart maintains data
-- Error recovery works
-- Performance acceptable
-
-## Migration Considerations
-
-- No migration needed (was non-functional demo)
-- Safe to remove all localStorage data
-- Users start fresh with file-based storage
-
-## Performance Requirements
-
-- List renders efficiently with 100+ personalities
-- No janky animations
-- Form interactions responsive
-- File operations non-blocking
-
-## Security Considerations
-
-- No sensitive data in localStorage
-- File permissions properly set
-- Input validation maintained
-- XSS prevention in place
-
-## Dependencies
-
-- **F-implement-user-interactions**: Requires all interactions working before final cleanup
